hotfix event ownerId

+++ src/http/controllers/event/event.controller.ts
@@ -9,6 +9,7 @@ import {
   Param,
   Put,
   Delete,
+  UnauthorizedException,
 } from '@nestjs/common'
 import {
   ApiBody,
@@ -98,6 +99,13 @@ export class EventController {
     @Request() req,
     @Body() updateEventDto: UpdateEventDTO,
   ) {
+    const ownerId = await this.eventService.getOwnerId(id)
+
+    if (req.user.id !== ownerId) {
+      throw new UnauthorizedException(
+        'You do not have permission to update this event',
+      )
+    }
     return this.eventService.update(id, updateEventDto, req.user)
   }
 
@@ -123,6 +131,13 @@ export class EventController {
   })
   @Delete(':id/delete')
   async delete(@Param('id') id: string, @Request() req) {
+    const ownerId = await this.eventService.getOwnerId(id)
+
+    if (req.user.id !== ownerId) {
+      throw new UnauthorizedException(
+        'You do not have permission to update this event',
+      )
+    }
     return this.eventService.delete(id, req.user)
   }
 
@@ -156,7 +171,6 @@ export class EventController {
     required: false,
     description: 'Number of items per page. Default: 10',
   })
-  @Get('all')
   @Get('all')
   findAll(@Query('page') page: number = 1, @Query('limit') limit: number = 10) {
     return this.eventService.findAll(Number(page) || 1, Number(limit) || 10)
